requirements.txt
streamlit
jsonschema
python-dotenv
requests
urllib3
